requirements.txt
torch
torchvision
transformers
diffusers
pandas
PyYAML
Pillow
numpy
tqdm
